+++ CHANGELOG.md
@@ -7,8 +7,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add new `LoudspeakerIllustration` React component.
 - Feature: Add new `Popover` React component.
 - Feature: Add new `CallToActionPopover` React component.
-- Fix: Vertically align buttons to the middle of the line.
+- Feature: Add `rowClassName` prop on `SideGrid` component.
+- Feature: Add `className`, `titleClassName` and `subtitleClassName` props
+  on `HeaderTitles`.
 - Feature: Add `quintuple` utilities in `k-utilities-margin`.
+- Fix: Vertically align buttons to the middle of the line.
 
 ## [5.7.0] - 2017-02-02